Update util func
